requirements.txt
torch>=1.10.0
torchvision>=0.11.1
einops
pytorch-lightning==1.6.4
torchmetrics==0.6.0
lightning-bolts>=0.5.0
tqdm
wandb
scipy
timm
scikit-learn
hydra-core
h5py
pandas
seaborn
umap-learn
open_clip_torch